Set compile warn level to W4: * code cleanup

+++ src/DetailPage.cpp
@@ -19,9 +19,6 @@ IMPLEMENT_DYNCREATE(CDetailPage, CPropertyPage)
 
 CDetailPage::CDetailPage() : CPropertyPage(CDetailPage::IDD)
 {
-	//{{AFX_DATA_INIT(CDetailPage)
-	//}}AFX_DATA_INIT
-
 	for (int i=0; i<LASTSERVICE; i++)
 	{
 		m_dlData[i] = _T("");
@@ -39,7 +36,6 @@ CDetailPage::~CDetailPage()
 void CDetailPage::DoDataExchange(CDataExchange* pDX)
 {
 	CPropertyPage::DoDataExchange(pDX);
-	//{{AFX_DATA_MAP(CDetailPage)
 	DDX_Control(pDX, IDC_PROGRESS_WEB_UL, m_ulCtrl[S_WEB]);
 	DDX_Control(pDX, IDC_PROGRESS_WEB_DL, m_dlCtrl[S_WEB]);
 	DDX_Control(pDX, IDC_PROGRESS_UDP_UL, m_ulCtrl[S_UDP]);
@@ -82,14 +78,11 @@ void CDetailPage::DoDataExchange(CDataExchange* pDX)
 	DDX_Text(pDX, IDC_WEB_TOTAL_DATA, m_totalData[S_WEB]);
 	DDX_Text(pDX, IDC_WEB_UL_DATA, m_ulData[S_WEB]);
 	DDX_Text(pDX, IDC_WEB_UL_SPEED, m_ulSpeed[S_WEB]);
-	//}}AFX_DATA_MAP
 }
 
 
 BEGIN_MESSAGE_MAP(CDetailPage, CPropertyPage)
-	//{{AFX_MSG_MAP(CDetailPage)
 	ON_WM_TIMER()
-	//}}AFX_MSG_MAP
 END_MESSAGE_MAP()
 
 /////////////////////////////////////////////////////////////////////////////
@@ -97,7 +90,6 @@ END_MESSAGE_MAP()
 
 void CDetailPage::OnTimer(UINT nIDEvent) 
 {
-	// TODO: Add your message handler code here and/or call default
 	if (nIDEvent == IDT_DETAIL)
 	{
 		DWORD64		lwDiff;
@@ -136,12 +128,10 @@ void CDetailPage::OnTimer(UINT nIDEvent)
 		SetTimer(IDT_DETAIL, 1000, NULL);
 
 	}
-//	CPropertyPage::OnTimer(nIDEvent);
 }
 
 BOOL CDetailPage::OnSetActive() 
 {
-	// TODO: Add your specialized code here and/or call the base class
 	int dl, ul;
 	dl = COptionsPage::GetDownloadSpeed();
 	ul = COptionsPage::GetUploadSpeed();
@@ -151,18 +141,17 @@ BOOL CDetailPage::OnSetActive()
 		m_dlCtrl[i].SetShowMaximum();
 		m_dlCtrl[i].SetMaximumTime(50);
 		m_dlCtrl[i].SetColor(COLORDOWN);
-		m_dlCtrl[i].SetRange(0,dl);
+		m_dlCtrl[i].SetRange(0,(short)dl);
 		m_ulCtrl[i].SetShowMaximum();
 		m_ulCtrl[i].SetMaximumTime(50);
 		m_ulCtrl[i].SetColor(COLORUP);
-		m_ulCtrl[i].SetRange(0, ul);
+		m_ulCtrl[i].SetRange(0, (short)ul);
 	}
 	return CPropertyPage::OnSetActive();
 }
 
 BOOL CDetailPage::OnKillActive() 
 {
-	// TODO: Add your specialized code here and/or call the base class
 	KillTimer(IDT_DETAIL);	
 	return CPropertyPage::OnKillActive();
 }
